requirements.txt
openai>=1.0.0
anthropic>=0.18.0
requests>=2.31.0
python-dotenv>=1.0.0
google-generativeai>=0.8.0,<1.0.0
rich>=13.7.0